+++ MemoryControl/Generics/RotationMatrix_4.cs
@@ -24,18 +24,18 @@ namespace CommonToolLib.Generics
 
         public RotationMatrix_4(Vector4_32 A, Vector4_32 B, Vector4_32 C)
         {
-            Matrix[0, 0] = (double)A.X;
-            Matrix[0, 1] = (double)A.Y;
-            Matrix[0, 2] = (double)A.Z;
-            Matrix[0, 3] = (double)A.W;
-            Matrix[1, 0] = (double)B.X;
-            Matrix[1, 1] = (double)B.Y;
-            Matrix[1, 2] = (double)B.Z;
-            Matrix[1, 3] = (double)B.W;
-            Matrix[2, 0] = (double)C.X;
-            Matrix[2, 1] = (double)C.Y;
-            Matrix[2, 2] = (double)C.Z;
-            Matrix[2, 3] = (double)C.W;
+            Matrix[0, 0] = A.X;
+            Matrix[0, 1] = A.Y;
+            Matrix[0, 2] = A.Z;
+            Matrix[0, 3] = A.W;
+            Matrix[1, 0] = B.X;
+            Matrix[1, 1] = B.Y;
+            Matrix[1, 2] = B.Z;
+            Matrix[1, 3] = B.W;
+            Matrix[2, 0] = C.X;
+            Matrix[2, 1] = C.Y;
+            Matrix[2, 2] = C.Z;
+            Matrix[2, 3] = C.W;
         }
         #endregion
