+++ MemoryControl/UI/Vector4Display.cs
@@ -116,10 +116,10 @@ namespace CommonToolLib.UI
             get => new CommonToolLib.Generics.Vector4_32(Convert.ToInt32(X), Convert.ToInt32(Y), Convert.ToInt32(Z), Convert.ToInt32(W));
             set
             {
-                X = (decimal)value.X;
-                Y = (decimal)value.Y;
-                Z = (decimal)value.Z;
-                W = (decimal)value.W;
+                X = value.X;
+                Y = value.Y;
+                Z = value.Z;
+                W = value.W;
             }
         }
